requirements.txt
Flask==3.1.2
tldextract==5.3.0
python-Levenshtein==0.27.3
gunicorn==23.0.0
